fix(D-03): keep corrupt-data UI within module gate
